refactor(session): ONE abort-first cancellation protocol + honest start-run wording
- The abort-first iterator wrapper (cancel the underlying work before
  delegating gen.return — the deadlock the cancellation door fixed)
  existed as three near-verbatim copies across session-remote.ts and
  invoke.ts; a boundary fix in one would silently drift the others.
  Now one helper (abortFirstIterator in collect.ts) with the protocol
  comment written once, three consumers.
- attach's steer fallback states intent, not fact: admission may still
  fail (lease held by a compaction or another run), so 'starting one;
  detaching will cancel it' promised two things that might never
  happen.
